67-rentals/requirements.txt
stripe
pycryptodome
mysql-connector-python

flask==3.0.0
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
geopy==2.3.0
